Removed the persistent scripting bridge app.
For some reason, when storing the scripting bridge app in an instance
variable, the app is not initialized properly. In this case, the number
of terminals that iTerm's bridge reported was always 0 until the
a terminal was manually created. Changing to a scenario where the bridge
is initialized on every command makes it work properly.